refactor: extract ChatterboxBase + api/chunked; drop _small from TTSEngine ABC
Three internal refactors from artifacts/analyses/stage-axis-audit-2026-05-20.md
that don't depend on lyra upstream work:

- New engines/_chatterbox_base.py mixin holds the verbatim _generate_chunked
  and the near-identical _generate_segmented. ChatterboxEngine (multilingual)
  overrides only a small _apply_segment_overrides hook for language_id
  dispatch; ChatterboxTurboEngine inherits the base as-is.
- New api_chunked.py hosts emit_chunk / generate_chunked / clone_chunked /
  write_done. api.py re-imports them under the historical underscore-prefixed
  names. api.py drops from 1125 → 939 LOC.
- ports/tts.py: drop `_small: bool = False` from the TTSEngine ABC. The
  attribute was Qwen-only and polluted every other engine. QwenEngine
  already owns it as an instance attribute (qwen.py:42). The four
  remaining `eng._small = True` mutation sites (api.py x2, daemon.py,
  daemon_protocol.py) use setattr so type-checkers don't complain about
  the (intentionally Qwen-only) attribute on a TTSEngine-typed binding.

Closes Findings 1, 2, 4 from the audit.

Refs #162
