+++ src/voicecli/api.py
@@ -12,6 +12,10 @@ from typing import TYPE_CHECKING
 if TYPE_CHECKING:
     from voicecli.ports.synthesis import SynthesisPort
 
+from voicecli.api_chunked import (
+    clone_chunked as _clone_chunked,
+    generate_chunked as _generate_chunked,
+)
 from voicecli.utils import OUTPUT_DIR, STT_OUTPUT_DIR, _Unrestricted
 
 log = logging.getLogger(__name__)
@@ -395,205 +399,8 @@ def _resolve_ref(ref: Path | str | None) -> Path:
 
 
 # ── Chunked output helpers ──────────────────────────────────────────────────
-
-
-def _emit_chunk(
-    eng,
-    method: str,
-    text: str,
-    voice,
-    out: Path,
-    index: int,
-    _total: int,
-    *,
-    mp3: bool = False,
-    daemon_fn=None,
-    **kwargs,
-) -> Path:
-    """Generate and save a single numbered chunk. Returns chunk path."""
-    stem = out.stem
-    chunk_path = out.parent / f"{stem}_{index:03d}.wav"
-    if daemon_fn is None or not daemon_fn(method, text, voice, chunk_path, **kwargs):
-        if method == "generate":
-            eng.generate(text, voice, chunk_path, **kwargs)
-        else:
-            eng.clone(
-                text,
-                kwargs.pop("ref_audio"),
-                chunk_path,
-                ref_text=kwargs.pop("ref_text", None),
-                **kwargs,
-            )
-    if mp3:
-        from voicecli.utils import wav_to_mp3
-
-        wav_to_mp3(chunk_path)
-    return chunk_path
-
-
-def _write_done(out: Path) -> Path:
-    done_path = out.with_suffix(".done")
-    done_path.write_text("done\n")
-    return done_path
-
-
-def _generate_chunked(
-    eng,
-    text,
-    voice,
-    out,
-    language,
-    extra_kwargs,
-    *,
-    chunk_size,
-    segments,
-    mp3=False,
-    daemon_fn=None,
-) -> list[Path]:
-    """Generate speech in chunks. Returns list of chunk paths."""
-    from voicecli.utils import smart_chunk
-
-    paths: list[Path] = []
-
-    if segments and len(segments) > 1:
-        total = len(segments)
-        for i, seg in enumerate(segments, 1):
-            kw = {**extra_kwargs, "language": seg.language or language}
-            if seg.instruct:
-                kw["instruct"] = seg.instruct
-            if seg.exaggeration is not None:
-                kw["exaggeration"] = seg.exaggeration
-            if seg.cfg_weight is not None:
-                kw["cfg_weight"] = seg.cfg_weight
-            if seg.flow_steps is not None:
-                kw["flow_steps"] = seg.flow_steps
-            if seg.cfg_alpha is not None:
-                kw["cfg_alpha"] = seg.cfg_alpha
-            if seg.temperature is not None:
-                kw["temperature"] = seg.temperature
-            if seg.top_p is not None:
-                kw["top_p"] = seg.top_p
-            if seg.min_p is not None:
-                kw["min_p"] = seg.min_p
-            if seg.repetition_penalty is not None:
-                kw["repetition_penalty"] = seg.repetition_penalty
-            seg_voice = seg.voice or voice
-            p = _emit_chunk(
-                eng,
-                "generate",
-                seg.text,
-                seg_voice,
-                out,
-                i,
-                total,
-                mp3=mp3,
-                daemon_fn=daemon_fn,
-                **kw,
-            )
-            paths.append(p)
-    else:
-        chunks = smart_chunk(text, chunk_size)
-        total = len(chunks)
-        for i, chunk_text in enumerate(chunks, 1):
-            p = _emit_chunk(
-                eng,
-                "generate",
-                chunk_text,
-                voice,
-                out,
-                i,
-                total,
-                mp3=mp3,
-                daemon_fn=daemon_fn,
-                language=language,
-                **extra_kwargs,
-            )
-            paths.append(p)
-
-    _write_done(out)
-    return paths
-
-
-def _clone_chunked(
-    eng,
-    text,
-    ref,
-    ref_text,
-    out,
-    language,
-    extra_kwargs,
-    *,
-    chunk_size,
-    segments,
-    mp3=False,
-    daemon_fn=None,
-) -> list[Path]:
-    """Clone voice in chunks. Returns list of chunk paths."""
-    from voicecli.utils import smart_chunk
-
-    paths: list[Path] = []
-
-    if segments and len(segments) > 1:
-        total = len(segments)
-        for i, seg in enumerate(segments, 1):
-            kw = {
-                **extra_kwargs,
-                "language": seg.language or language,
-                "ref_audio": ref,
-                "ref_text": ref_text,
-            }
-            if seg.exaggeration is not None:
-                kw["exaggeration"] = seg.exaggeration
-            if seg.cfg_weight is not None:
-                kw["cfg_weight"] = seg.cfg_weight
-            if seg.flow_steps is not None:
-                kw["flow_steps"] = seg.flow_steps
-            if seg.cfg_alpha is not None:
-                kw["cfg_alpha"] = seg.cfg_alpha
-            if seg.temperature is not None:
-                kw["temperature"] = seg.temperature
-            if seg.top_p is not None:
-                kw["top_p"] = seg.top_p
-            if seg.min_p is not None:
-                kw["min_p"] = seg.min_p
-            if seg.repetition_penalty is not None:
-                kw["repetition_penalty"] = seg.repetition_penalty
-            p = _emit_chunk(
-                eng,
-                "clone",
-                seg.text,
-                None,
-                out,
-                i,
-                total,
-                mp3=mp3,
-                daemon_fn=daemon_fn,
-                **kw,
-            )
-            paths.append(p)
-    else:
-        chunks = smart_chunk(text, chunk_size)
-        total = len(chunks)
-        for i, chunk_text in enumerate(chunks, 1):
-            p = _emit_chunk(
-                eng,
-                "clone",
-                chunk_text,
-                None,
-                out,
-                i,
-                total,
-                mp3=mp3,
-                daemon_fn=daemon_fn,
-                language=language,
-                ref_audio=ref,
-                ref_text=ref_text,
-                **extra_kwargs,
-            )
-            paths.append(p)
-
-    _write_done(out)
-    return paths
+# Implementation lives in voicecli.api_chunked; imported at module top under
+# the historical underscore-prefixed names that the rest of api.py uses.
 
 
 # ── Public API ───────────────────────────────────────────────────────────────
@@ -724,7 +531,9 @@ def generate(
 
             eng = get_engine(r_engine)
         if r_fast and r_engine in QWEN_ENGINES:
-            eng._small = True
+            # `_small` is a QwenEngine-only knob; the QWEN_ENGINES guard
+            # narrows eng to QwenEngine at runtime. setattr avoids ABC noise.
+            setattr(eng, "_small", True)
         chunk_paths = _generate_chunked(
             eng,
             r_text,
@@ -889,7 +698,9 @@ def clone(
 
             eng = get_engine(r_engine)
         if r_fast and r_engine in QWEN_ENGINES:
-            eng._small = True
+            # `_small` is a QwenEngine-only knob; the QWEN_ENGINES guard
+            # narrows eng to QwenEngine at runtime. setattr avoids ABC noise.
+            setattr(eng, "_small", True)
         chunk_paths = _clone_chunked(
             eng,
             r_text,
