+++ src/voicecli/engines/chatterbox.py
@@ -1,20 +1,20 @@
 # pyright: ignore — excluded from pyrightconfig.json (heavy ML deps, no type stubs)
 from __future__ import annotations
 
-import numpy as np
 import soundfile as sf
 from pathlib import Path
 from typing import TYPE_CHECKING
 
-from voicecli.engine import TTSEngine, cuda_guard
+from voicecli.engine import cuda_guard
+from voicecli.engines._chatterbox_base import ChatterboxBase
 from voicecli.models import CHATTERBOX_MODEL, warn_if_first_download
-from voicecli.utils import resolve_language as _resolve_language, split_sentences
+from voicecli.utils import resolve_language as _resolve_language
 
 if TYPE_CHECKING:
     from voicecli.markdown import Segment
 
 
-class ChatterboxEngine(TTSEngine):
+class ChatterboxEngine(ChatterboxBase):
     name = "chatterbox"
 
     def __init__(self):
@@ -37,57 +37,9 @@ class ChatterboxEngine(TTSEngine):
                 print("[chatterbox] Model loaded.")
         return self._model
 
-    def _generate_chunked(self, text: str, **gen_kwargs) -> np.ndarray:
-        """Generate audio in sentence-sized chunks and concatenate."""
-        model = self._load_model()
-        chunks = split_sentences(text)
-        wavs = []
-        for i, chunk in enumerate(chunks):
-            print(f"  [{i + 1}/{len(chunks)}] {chunk[:60]}...")
-            kw = {**gen_kwargs, "text": chunk}
-            wav = model.generate(**kw)
-            wavs.append(wav.squeeze().cpu().numpy())
-        return np.concatenate(wavs)
-
-    def _generate_segmented(
-        self,
-        segments: list[Segment],
-        base_kwargs: dict,
-        default_gap: int = 0,
-        default_crossfade: int = 0,
-    ) -> np.ndarray:
-        """Generate audio per-segment with individual overrides, then concatenate."""
-        from voicecli.utils import concat_audio
-
-        all_wavs: list[np.ndarray] = []
-        for i, seg in enumerate(segments):
-            print(f"  [{i + 1}/{len(segments)}] {seg.text[:60]}...")
-            kw = {**base_kwargs}
-            if seg.exaggeration is not None:
-                kw["exaggeration"] = seg.exaggeration
-            if seg.cfg_weight is not None:
-                kw["cfg_weight"] = seg.cfg_weight
-            if seg.temperature is not None:
-                kw["temperature"] = seg.temperature
-            if seg.top_p is not None:
-                kw["top_p"] = seg.top_p
-            if seg.min_p is not None:
-                kw["min_p"] = seg.min_p
-            if seg.repetition_penalty is not None:
-                kw["repetition_penalty"] = seg.repetition_penalty
-            if seg.language is not None:
-                kw["language_id"] = _resolve_language(seg.language)
-            audio = self._generate_chunked(seg.text, **kw)
-            all_wavs.append(audio)
-
-        gaps = [
-            seg.segment_gap if seg.segment_gap is not None else default_gap for seg in segments[1:]
-        ]
-        xfades = [
-            seg.crossfade if seg.crossfade is not None else default_crossfade
-            for seg in segments[1:]
-        ]
-        return concat_audio(all_wavs, self._load_model().sr, gaps, xfades)
+    def _apply_segment_overrides(self, kw: dict, seg: Segment) -> None:
+        if seg.language is not None:
+            kw["language_id"] = _resolve_language(seg.language)
 
     def generate(self, text: str, voice: str | None, output_path: Path, **kwargs) -> Path:
         language = _resolve_language(kwargs.get("language", "English"))
